move artifact management from subt/main.py to subt/artf_filter.py
Note, that artifacts are no longer reported at the return home!
